moves to system check, fixes issue with not being a relative path
