+++ src/documents/tests/test_api_objects.py
@@ -239,7 +239,7 @@ class TestApiStoragePaths(DirectoriesMixin, APITestCase):
                     "/{created_year_short}/{created_month}/{created_month_name}"
                     "/{created_month_name_short}/{created_day}/{added}/{added_year}"
                     "/{added_year_short}/{added_month}/{added_month_name}"
-                    "/{added_month_name_short}/{added_day}/{asn}/{tags}"
+                    "/{added_month_name_short}/{added_day}/{asn}"
                     "/{tag_list}/{owner_username}/{original_name}/{doc_pk}/",
                 },
             ),
