+++ src/documents/tests/test_checks.py
@@ -6,6 +6,7 @@ from django.test import TestCase
 from django.test import override_settings
 
 from documents.checks import changed_password_check
+from documents.checks import filename_format_check
 from documents.checks import parser_check
 from documents.models import Document
 from documents.tests.factories import DocumentFactory
@@ -70,6 +71,20 @@ class TestDocumentChecks(TestCase):
                     Error(
                         "No parsers found. This is a bug. The consumer won't be "
                         "able to consume any documents without parsers.",
+                    ),
+                ],
+            )
+
+    def test_filename_format_check(self):
+        self.assertEqual(filename_format_check(None), [])
+
+        with override_settings(FILENAME_FORMAT="{created}/{title}"):
+            self.assertEqual(
+                filename_format_check(None),
+                [
+                    Error(
+                        "Filename format {created}/{title} is using the old style, please update to use double curly brackets",
+                        hint="{{ created }}/{{ title }}",
                     ),
                 ],
             )
